beacon: correct W param for the beacon proposal threshold (#3845)
## Motivation
<!-- Please mention the issue fixed by this PR or detailed motivation -->
part of #3685
<!-- `Closes #XXXX, closes #XXXX, ...` links mentioned issues to this PR and automatically closes them when this it's merged -->

## Changes
<!-- Please describe in detail the changes made -->
- add systest for beacon uniqueness across epochs
- add multiple nodes unit tests
- correct W param to be num of ATXs instead of total epoch weight

+++ cmd/root.go
@@ -154,7 +154,7 @@ func AddCommands(cmd *cobra.Command) {
 
 	/**======================== Beacon Flags ========================== **/
 
-	cmd.PersistentFlags().Uint64Var(&cfg.Beacon.Kappa, "beacon-kappa",
+	cmd.PersistentFlags().IntVar(&cfg.Beacon.Kappa, "beacon-kappa",
 		cfg.Beacon.Kappa, "Security parameter (for calculating ATX threshold)")
 	cmd.PersistentFlags().Var((*types.RatVar)(cfg.Beacon.Q), "beacon-q",
 		"Ratio of dishonest spacetime (for calculating ATX threshold). It should be a string representing a rational number.")
